CardManager pain

+++ src/dolphin/card/CARDStat.c
@@ -147,9 +147,9 @@ s32 CARDSetStatusAsync(s32 chan, s32 fileNo, CARDStat* stat,
 	return result;
 }
 
-long CARDSetStatus(long chan, long fileNo, struct CARDStat* stat)
+s32 CARDSetStatus(s32 chan, s32 fileNo, struct CARDStat* stat)
 {
-	long result = CARDSetStatusAsync(chan, fileNo, stat, __CARDSyncCallback);
+	s32 result = CARDSetStatusAsync(chan, fileNo, stat, __CARDSyncCallback);
 
 	if (result < 0) {
 		return result;
